pyproject.toml
[project]
name = "meshtastic-hexagonal"
version = "0.0.2"
requires-python = ">=3.12.0"
dependencies = [
    "meshtastic>=2.7.7",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.12.5",
    "textual>=0.40.0",
    "flask[async]>=3.1.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.14.14",
]

[project.scripts]
meshtastic-hexagonal = "meshcore.main:main"
meshcore-tui = "meshcore.tui_main:main"
meshcore-web = "meshcore.web_main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
